Move TransactionContent and BlockContext
