+++ src/dfi/tokens.cpp
@@ -9,6 +9,7 @@
 #include <chainparams.h>  // Params()
 #include <core_io.h>
 #include <dfi/evm.h>
+#include <dfi/mn_checks.h>
 #include <ffi/cxx.h>
 #include <ffi/ffihelpers.h>
 #include <primitives/transaction.h>
@@ -69,7 +70,7 @@ Res CTokensView::CreateDFIToken() {
 
 ResVal<DCT_ID> CTokensView::CreateToken(const CTokensView::CTokenImpl &token,
                                         bool isPreBayfront,
-                                        BlockContext blockCtx) {
+                                        BlockContext *blockCtx) {
     if (GetTokenByCreationTx(token.creationTx)) {
         return Res::Err("token with creation tx %s already exists!", token.creationTx.ToString());
     }
@@ -101,18 +102,20 @@ ResVal<DCT_ID> CTokensView::CreateToken(const CTokensView::CTokenImpl &token,
                       id.ToString().c_str());
         }
 
-        const auto shouldCreateDst20 = blockCtx.GetEVMEnabledForBlock();
-        const auto &evmTemplate = blockCtx.GetEVMTemplate();
-        if (shouldCreateDst20 && evmTemplate) {
-            CrossBoundaryResult result;
-            evm_try_unsafe_create_dst20(result,
-                                        evmTemplate->GetTemplate(),
-                                        token.creationTx.GetHex(),
-                                        rust::string(token.name.c_str()),
-                                        rust::string(token.symbol.c_str()),
-                                        id.v);
-            if (!result.ok) {
-                return Res::Err("Error creating DST20 token: %s", result.reason);
+        if (blockCtx) {
+            const auto shouldCreateDst20 = blockCtx->GetEVMEnabledForBlock();
+            const auto &evmTemplate = blockCtx->GetEVMTemplate();
+            if (shouldCreateDst20 && evmTemplate) {
+                CrossBoundaryResult result;
+                evm_try_unsafe_create_dst20(result,
+                                            evmTemplate->GetTemplate(),
+                                            token.creationTx.GetHex(),
+                                            rust::string(token.name.c_str()),
+                                            rust::string(token.symbol.c_str()),
+                                            id.v);
+                if (!result.ok) {
+                    return Res::Err("Error creating DST20 token: %s", result.reason);
+                }
             }
         }
     } else {
